refactor: 키보드 툴바 헬퍼 클래스 구현
