+++ FirstSeminar/FirstSeminar/Sources/Presentaion/ViewController.swift
@@ -112,37 +112,12 @@ final class ViewController: BaseViewController {
     }
 
     private func configureToolbar() {
-        let toolbar = UIToolbar()
-        toolbar.sizeToFit()
-
-        let previousButton = UIBarButtonItem(
-            image: UIImage(systemName: "chevron.up"),
-            style: .plain,
+        let toolbar = KeyboardToolbarHelper.createToolbar(
             target: self,
-            action: #selector(previousField)
+            previousSelector: #selector(previousField),
+            nextSelector: #selector(nextField),
+            dismissSelector: #selector(dismissKeyboard)
         )
-
-        let nextButton = UIBarButtonItem(
-            image: UIImage(systemName: "chevron.down"),
-            style: .plain,
-            target: self,
-            action: #selector(nextField)
-        )
-
-        let flexSpace = UIBarButtonItem(
-            barButtonSystemItem: .flexibleSpace,
-            target: nil,
-            action: nil
-        )
-
-        let doneButton = UIBarButtonItem(
-            title: "키보드 내리기",
-            style: .plain,
-            target: self,
-            action: #selector(dismissKeyboard)
-        )
-
-        toolbar.items = [previousButton, nextButton, flexSpace, doneButton]
 
         titleTextField.inputAccessoryView = toolbar
         contentTextView.inputAccessoryView = toolbar
